Added hash_ring repository into the repo.
Taken from andreweduffy/hash_ring and placed inside of here.
The original version of hash_ring did not have an explicit
license in the repository, meaning we may need to take extra
steps to legitimize this in the future, but for now I don't
think anyone will protest to this being shipped with CodaLab.