MultiDiskBundleStore creates a default partition that points to CODALAB_HOME
This makes transitioning from the old SingleDiskBundleStore over
to MultiDiskBundleStore zero-effort. The CODALAB_HOME/partitions
directory is created, and then CODALAB_HOME/partitions/default
is setup to link to CODALAB_HOME, which means that it will use
the existing CODALAB_HOME/bundles and CODALAB_HOME/temp as the
bundle and temp directories.