Added checking/computing of data_hash for health check
Now, if a bundle has a NoneType data_hash, the correct hash will
be calculated, and subsequently committed back to the DB if the
--force flag is passed. Also added a new flag, --data-hash
(or -d short form) that will check every bundle to recalculate
data_hash for every bundle and compare against what is in the
database.

Also added --repair flag to fix up the data_hash for bundles
whose data_hash stored in the database does not reflect the
disk contents.