Tell user to click Upload button
When we receive a bare `cl upload` command on the web terminal.
We do this browsers do not allow JavaScript code to directly
open a file dialog outside of a user-initiated event. Because
there is a network call in the chain of events from typing in
`cl upload` in the web terminal to receiving an Upload "UI action"
from the server, the browser considers the final callback to be
script-initiated, not directly user-initiated.

Resolves #205