Bugfix: Offline bundles should be removed from worker_run
Previously, the only way an offline row in worker_run can be removed is
when the worker that is assigned to it is removed (as a dead worker).

If the worker never gets detected as 'dead' by the server for whatever
reason then the offline row stays in worker_run forever.

Also there was a bug in resume_bundle() where the last_updated timestamp
is not updated if the bundle is in RUNNING state. This field should be
regularly updated so that last_updated doesn't fall too far behind and
get moved to WORKER_OFFLINE

Fixes #876