Call resume_bundle before calling finalize_bundle
When a bundle is being finalized, its metadata.run_status is not
updated to reflect that it is "Finalizing Bundle", in fact leaving
the run_status to be whatever it was the last time resume_bundle
was called.

This forever leaves run_status as typically one of "Upload started"
or "Uploading results: yyy done (archived size)", neither of which
look good and worse, may confuse an inexperienced user.

Solution is to update the bundle metadata using resume_bundle when
finalizing it, and update metadata again during finish_bundle.