Change name of general worker docker network
Previously the name of the general worker docker network that contains
all the containers was a prefix of the names of the specific internal
and external networks that determine the network connection status.

This caused problems when workers were started on machines with these
networks already existing. The Docker SDK's network search method
returns networks for which the search term is a substring (and in
apparent random order from call to call), so it was
possible for us to pick one of the external or internal networks when
trying to reuse the general network.

In that situation we would try to connect the container to the same
network twice, leading to the annoying endpoint already exists errors.
With this PR, no network name should be a substring of another, so even
when we reuse networks, we should not pick the wrong network.