Allow a system to configure worker-manager specific values
Before this PR, configurable values specific to a worker manager type
like AWS Batch had to be hardcoded in the respective worker manager
files.

This PR allows worker manager types to define their own config
structures, and the main worker manager wrapper only takes in a filepath
and passes it on to them for them to parse as they see fit.

Another way to do this might have been defining parsers in the specific
worker manager subclasses, then importing these parsers in the wrapper
class and adding them as subcommands to the worker-manager command.

That would enable specific worker manager types to define their own
command line arguments without having to modify the wrapper class, but
it would be a bit more complicated to code compared to offloading
everything to JSON files.