Fix bundle manager bug for bad queue argument (#2045)
* Fix bundle manager edge case crash

Small reference bug. If there's a run where `request_queue` is set but
the value is not `tag=<something>` the current code hits a buggy path.

This is a hotfix, but I think the whole `request_queue` semantics need
an update. As things as are `request_queue` can ONLY be used with
`--request_queue tag=<something>` and causes bugs if specified in any
other shape.