Upgrade urllib3 to 1.26.4 (#3359)
* Upgrade urllib3 to 1.26.4

Fixes a security vulnerability (https://github.com/advisories/GHSA-5phf-pp7p-vc2r). Technically, this vulnerability doesn't affect urllib3 < 1.26 (so it doesn't impact codalab; seems to be an issue with github showing a false positive), but we might as well upgrade the package so that github doesn't keep showing this alert as a false positive.

* Bump boto3

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>